# A bit (~5%) faster Rust