--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,2 @@
+src/rust/target
+src/rust/Cargo.lock