--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
 src/rust/target
 src/rust/Cargo.lock
+prime