*------- assembly (nasm) ---------*
NASM version 2.16.01 compiled on Mar 18 2023
602489
2289

*---- c native (opt-level=3) -----*
gcc (GCC) 12.2.1 20230201
602489
2617ms

*--- c++ native (opt-level=3) ----*
g++ (GCC) 12.2.1 20230201
602489
2297ms

*--- rust native (opt-level=3) ---*
rustc 1.69.0 (84c898d65 2023-04-16)
    Finished release [optimized] target(s) in 0.00s
602489
2375ms

*-------------- go ---------------*
go version go1.20.3 linux/amd64
602489
2630ms