@ali77gh 2023-05-01
Too match complexity for 5%.
All other languages handles with single source code file without any other config files and stuff.
What do you thing? should I merge it?